Updated readme, installation script, login and sync scripts for better workflow. Does not get locked out because of failure of periodical sync, now it retries. Login script lockout still requires manual labour.

+++ rclone-sync.sh
@@ -4,13 +4,7 @@
 FLAGPATH=$HOME/.config/rclone
 DRIVEPATH=$HOME/GoogleDrive
 
-# Check if the flag file from the last session exists; if it does, don't proceed with sync
-if [ -e "$FLAGPATH"/last_session_sync_failed ]; then
-  echo "Last session sync failed. Skipping the cron job."
-  exit 0
-fi
-
-# Check if the flag file exists; if it does, don't proceed with sync
+# Check if the startup sync failure flag file exists; if it does, don't proceed with sync
 if [ -e "$FLAGPATH"/no_sync_flag ]; then
   echo "Startup sync skipped due to a previous failure. Skipping the cron job."
   exit 0
@@ -19,9 +13,17 @@ fi
 # Perform the sync from local machine to Google Drive
 if ! rclone sync -v "$DRIVEPATH" GoogleDrive: > "$FLAGPATH"/sync-log.log 2>&1;
 then
-  echo "Error during sync to Google Drive. Please check and resolve the issue."
-  touch "$FLAGPATH"/last_session_sync_failed  # Create a flag file to indicate last session sync failure
+ echo "Error during sync to Google Drive. Please check and resolve the issue."
+
+ touch "$FLAGPATH"/last_session_sync_failed  # Create a flag file to indicate last session sync failure
+ notify-send 'Google Drive Sync' 'Periodic sync has failed.' --icon=dialog-error
+
 else
   echo "Sync to Google Drive successful."
-  rm -f "$FLAGPATH"/last_session_sync_failed  # Remove the flag file if sync is successful
+  if [ -e "$FLAGPATH"/last_session_sync_failed ]; then
+
+    rm -f "$FLAGPATH"/last_session_sync_failed  # Remove the flag file if sync is successful
+    notify-send 'Google Drive Sync' 'Periodic sync has resumed working.' --icon=dialog-information
+
+  fi
 fi
